fix block_store_test and ledger_store_test bug (#696)
