+++ core/store/ledgerstore/ledger_store_test.go
@@ -20,15 +20,13 @@ package ledgerstore
 
 import (
 	"fmt"
-	"os"
-	"testing"
-	"time"
-
 	"github.com/ontio/dad-go-crypto/keypair"
 	"github.com/ontio/dad-go/account"
-	"github.com/ontio/dad-go/common"
+	"github.com/ontio/dad-go/common/config"
 	"github.com/ontio/dad-go/common/log"
-	"github.com/ontio/dad-go/core/types"
+	"github.com/ontio/dad-go/core/genesis"
+	"os"
+	"testing"
 )
 
 var testBlockStore *BlockStore
@@ -87,26 +85,31 @@ func TestInitLedgerStoreWithGenesisBlock(t *testing.T) {
 	acc2 := account.NewAccount("")
 	acc3 := account.NewAccount("")
 	acc4 := account.NewAccount("")
+	acc5 := account.NewAccount("")
+	acc6 := account.NewAccount("")
+	acc7 := account.NewAccount("")
 
-	bookkeepers := []keypair.PublicKey{acc1.PublicKey, acc2.PublicKey, acc3.PublicKey, acc4.PublicKey}
-	bookkeeper, err := types.AddressFromBookkeepers(bookkeepers)
-	if err != nil {
-		t.Errorf("AddressFromBookkeepers error %s", err)
-		return
-	}
-	header := &types.Header{
-		Version:          123,
-		PrevBlockHash:    common.Uint256{},
-		TransactionsRoot: common.Uint256{},
-		Timestamp:        uint32(uint32(time.Date(2017, time.February, 23, 0, 0, 0, 0, time.UTC).Unix())),
-		Height:           uint32(0),
-		ConsensusData:    1234567890,
-		NextBookkeeper:   bookkeeper,
-	}
-	block := &types.Block{
-		Header:       header,
-		Transactions: []*types.Transaction{},
-	}
+	bookkeepers := []keypair.PublicKey{acc1.PublicKey, acc2.PublicKey, acc3.PublicKey, acc4.PublicKey, acc5.PublicKey, acc6.PublicKey, acc7.PublicKey}
+	//bookkeeper, err := types.AddressFromBookkeepers(bookkeepers)
+	//if err != nil {
+	//	t.Errorf("AddressFromBookkeepers error %s", err)
+	//	return
+	//}
+	genesisConfig := config.DefConfig.Genesis
+	block, err := genesis.BuildGenesisBlock(bookkeepers, genesisConfig)
+	//header := &types.Header{
+	//	Version:          123,
+	//	PrevBlockHash:    common.Uint256{},
+	//	TransactionsRoot: common.Uint256{},
+	//	Timestamp:        uint32(uint32(time.Date(2017, time.February, 23, 0, 0, 0, 0, time.UTC).Unix())),
+	//	Height:           uint32(0),
+	//	ConsensusData:    1234567890,
+	//	NextBookkeeper:   bookkeeper,
+	//}
+	//block := &types.Block{
+	//	Header:       header,
+	//	Transactions: []*types.Transaction{},
+	//}
 
 	err = testLedgerStore.InitLedgerStoreWithGenesisBlock(block, bookkeepers)
 	if err != nil {
